CLTK-36 : getClosetEntity 수정

+++ src/main/java/team/closetalk/closet/service/ClosetItemService.java
@@ -3,6 +3,7 @@ package team.closetalk.closet.service;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 import team.closetalk.closet.dto.ClosetItemDto;
@@ -10,7 +11,9 @@ import team.closetalk.closet.entity.ClosetEntity;
 import team.closetalk.closet.entity.ClosetItemEntity;
 import team.closetalk.closet.repository.ClosetItemRepository;
 import team.closetalk.closet.repository.ClosetRepository;
+import team.closetalk.user.entity.UserEntity;
 
+import java.util.List;
 import java.util.Map;
 
 @Slf4j
@@ -22,36 +25,56 @@ public class ClosetItemService {
     private final EntityRetrievalService entityRetrievalService;
 
     // 1. 해당 옷장의 단일 아이템 조회
-    public ClosetItemDto readClosetItem(Long itemId) {
-        ClosetItemEntity item = getClosetItemEntity(itemId);
+    public ClosetItemDto readClosetItem(Long itemId, Authentication authentication) {
+        String nickName = getUserEntity(authentication.getName()).getNickname();
+        ClosetItemEntity item = getClosetItemEntity(itemId, nickName);
 
         log.info("[{}]의 [{}]번 아이템 조회 완료",
                 item.getClosetId().getClosetName(), item.getId());
         return ClosetItemDto.toClosetItemDto(item);
     }
 
-    public void modifyClosetItem(Long itemId, Map<String, String> itemParams) {
-        ClosetItemEntity item = getClosetItemEntity(itemId);
+    public void modifyClosetItem(Long itemId, Map<String, String> itemParams,
+                                 Authentication authentication) {
+        String nickName = getUserEntity(authentication.getName()).getNickname();
+        ClosetItemEntity item = getClosetItemEntity(itemId, nickName);
 
-        ClosetEntity closet = closetRepository
-                .findById(Long.valueOf(itemParams.get("closetId")))
-                .orElseThrow(() -> {
-                    log.error("존재하지 않는 closet_id : {}", itemParams.get("closetId"));
-                    return new ResponseStatusException(HttpStatus.NOT_FOUND);
-                });
-
-        closetItemRepository.save(item.updateEntity(itemParams, closet));
-        log.info("[{}]번 아이템 수정 완료", itemId);
+        if (itemParams.get("changeCloset").isEmpty()) {
+            log.error("존재하지 않는 closet_name : {}", itemParams.get("changeCloset"));
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+        else {
+            ClosetEntity closet = getClosetEntity(itemParams.get("changeCloset"), nickName);
+            closetItemRepository.save(item.updateEntity(itemParams, closet));
+            log.info("[{}]번 아이템 수정 완료", itemId);
+        }
     }
 
-    public void deleteClosetItem(Long itemId) {
-        getClosetItemEntity(itemId); // 해당 아이템이 존재하는 지 확인(굳이 안해도 됨)
-        log.info("[{}]번 아이템 삭제 완료", itemId);
-        closetItemRepository.deleteById(itemId);
+    public void deleteClosetItem(Long itemId, Authentication authentication) {
+        String nickName = getUserEntity(authentication.getName()).getNickname();
+        ClosetEntity closetEntity = getClosetItemEntity(itemId, nickName).getClosetId();
+
+        if (closetRepository.existsByIdAndUserId_LoginId(closetEntity.getId(), authentication.getName())) {
+            closetItemRepository.deleteById(itemId);
+            log.info("[{}]번 아이템 삭제 완료", itemId);
+        } else {
+            log.error("[{}]번 아이템이 존재하지 않음", itemId);
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
+    }
+
+    // closetId로 해당 ClosetEntity 찾기
+    private ClosetEntity getClosetEntity(String closetName, String nickName) {
+        return entityRetrievalService.getClosetEntity(closetName, nickName);
     }
 
     // itemId로 해당 ClosetItemEntity 찾기
-    private ClosetItemEntity getClosetItemEntity(Long itemId) {
-        return entityRetrievalService.getClosetItemEntity(itemId);
+    private ClosetItemEntity getClosetItemEntity(Long itemId, String nickName) {
+        return entityRetrievalService.getClosetItemEntity(itemId, nickName);
+    }
+
+    // LoginId == authentication.getName() 사용자 찾기
+    private UserEntity getUserEntity(String LoginId) {
+        return entityRetrievalService.getUserEntity(LoginId);
     }
 }
